:bug: Temporarely moved ability guard

+++ back-body-blast/src/modules/admin/workout/admin-workout.controller.ts
@@ -8,8 +8,8 @@ import {
   Patch,
   Post,
   Query,
+  Req,
   UseFilters,
-  UseGuards,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
@@ -19,74 +19,85 @@ import { AppResponses } from '../../../decorators/app-responses.decorator';
 import { AppSingleResponse } from '../../../dto/app-single-response.dto';
 import { AppStatusResponse } from '../../../dto/app-status-response.dto';
 import { AppPagination } from '../../../utils/app-pagination.util';
-import { UserRole } from '../../../constants/constants';
 import { MainExceptionFilter } from '../../../exceptions/main-exception.filter';
-import { RoleGuard } from '../../authentication/guards/role.guard';
 import { AdminWorkoutService } from './admin-workout.service';
 import { CreateWorkoutByAdminRequest } from './dto/admin-create-wrokout.dto';
 import { GetWorkoutByAdminDTO, GetWorkoutForUserByAdminRequest } from './dto/admin-get-workout.dto';
 import { UpdateWorkoutByAdminRequest } from './dto/admin-update-workout.dto';
 import { WorkoutEntity } from '../../../modules/core/workout/entity/workout.entity';
 import { AppDatePagination } from '../../../utils/app-date-pagination.util';
-import { Action } from '../../../modules/ability/ability.factory';
-import { CheckAbilities } from '../../../decorators/ability.decorator';
-import { AbilityGuard } from '../../../modules/authentication/guards/ability.guard';
 import { AppAuthGuard } from '../../../modules/authentication/guards/appAuth.guard';
+import { AbilityFactory, Action } from '../../../modules/ability/ability.factory';
+import { RequestWithUser } from '../../../modules/authentication/types/requestWithUser.type';
+import { MainException } from '../../../exceptions/main.exception';
 
 @Controller('admin/workouts')
 @ApiTags('Admin workouts')
 @UseFilters(MainExceptionFilter)
 @UsePipes(ValidationPipe)
-@AppAuthGuard(RoleGuard(UserRole.Admin))
+@AppAuthGuard()
 export class AdminWorkoutController {
-  constructor(private readonly adminService: AdminWorkoutService) {}
+  constructor(
+    private readonly adminService: AdminWorkoutService,
+    private readonly abilityFactory: AbilityFactory,
+  ) {}
 
   @Post()
-  @UseGuards(AbilityGuard)
-  @CheckAbilities({ action: Action.Create, subject: WorkoutEntity })
   @AppResponses({ status: 201, type: AppSingleResponse.type(AppSingleResponse) })
   @Throttle(5, 1)
-  async create(@Body() request: CreateWorkoutByAdminRequest) {
+  async create(@Req() req: RequestWithUser, @Body() request: CreateWorkoutByAdminRequest) {
+    const ability = this.abilityFactory.defineAbility(req.user);
+
+    if (!ability.can(Action.Manage, WorkoutEntity)) throw MainException.forbidden('Cannot create workout');
     return await this.adminService.create(request);
   }
 
   @Get()
-  @UseGuards(AbilityGuard)
-  @CheckAbilities({ action: Action.Read, subject: WorkoutEntity })
   @AppResponses({ status: 200, type: AppPagination.Response.type(WorkoutEntity) })
-  async getAll(@Query() query: AppPagination.Request) {
+  async getAll(@Req() req: RequestWithUser, @Query() query: AppPagination.Request) {
+    const ability = this.abilityFactory.defineAbility(req.user);
+
+    if (!ability.can(Action.Manage, WorkoutEntity)) throw MainException.forbidden('Cannot get workout');
     return await this.adminService.findAll(query);
   }
 
   @Get('date')
-  @UseGuards(AbilityGuard)
-  @CheckAbilities({ action: Action.Read, subject: WorkoutEntity })
   @AppResponses({ status: 200, type: AppDatePagination.Response.type(WorkoutEntity) })
-  async getAllByDate(@Query() query: GetWorkoutForUserByAdminRequest) {
+  async getAllByDate(@Req() req: RequestWithUser, @Query() query: GetWorkoutForUserByAdminRequest) {
+    const ability = this.abilityFactory.defineAbility(req.user);
+
+    if (!ability.can(Action.Manage, WorkoutEntity)) throw MainException.forbidden('Cannot get workout');
     return await this.adminService.findAllByDate(query);
   }
 
   @Get(':id')
-  @UseGuards(AbilityGuard)
-  @CheckAbilities({ action: Action.Read, subject: WorkoutEntity })
   @AppResponses({ status: 200, type: AppSingleResponse.type(GetWorkoutByAdminDTO) })
-  async getOne(@Param('id', ParseUUIDPipe) id: string) {
+  async getOne(@Req() req: RequestWithUser, @Param('id', ParseUUIDPipe) id: string) {
+    const ability = this.abilityFactory.defineAbility(req.user);
+
+    if (!ability.can(Action.Manage, WorkoutEntity)) throw MainException.forbidden('Cannot get workout');
     return await this.adminService.findOne(id);
   }
 
   @Patch(':id')
-  @UseGuards(AbilityGuard)
-  @CheckAbilities({ action: Action.Update, subject: WorkoutEntity })
   @AppResponses({ status: 200, type: AppSingleResponse.type(AppSingleResponse) })
-  async update(@Param('id', ParseUUIDPipe) id: string, @Body() body: UpdateWorkoutByAdminRequest) {
+  async update(
+    @Req() req: RequestWithUser,
+    @Param('id', ParseUUIDPipe) id: string,
+    @Body() body: UpdateWorkoutByAdminRequest,
+  ) {
+    const ability = this.abilityFactory.defineAbility(req.user);
+
+    if (!ability.can(Action.Manage, WorkoutEntity)) throw MainException.forbidden('Cannot update workout');
     return await this.adminService.update(id, body);
   }
 
   @Delete(':id')
-  @UseGuards(AbilityGuard)
-  @CheckAbilities({ action: Action.Delete, subject: WorkoutEntity })
   @AppResponses({ status: 200, type: AppSingleResponse.type(AppStatusResponse) })
-  async deleteOne(@Param('id', ParseUUIDPipe) id: string) {
+  async deleteOne(@Req() req: RequestWithUser, @Param('id', ParseUUIDPipe) id: string) {
+    const ability = this.abilityFactory.defineAbility(req.user);
+
+    if (!ability.can(Action.Manage, WorkoutEntity)) throw MainException.forbidden('Cannot delete workout');
     return await this.adminService.deleteOne(id);
   }
 }
